requirements.txt
openai>=1.60.0
pydantic>=2.7.0
PyYAML>=6.0.1
APScheduler>=3.10.4
pypdf>=4.2.0
python-docx>=1.1.2
beautifulsoup4>=4.12.3
requests>=2.32.3
genanki>=0.13.1
Pillow>=10.4.0
pytesseract>=0.3.10
python-dotenv>=1.0.1
PyMuPDF>=1.24.9
